nit fixes to utils file
